pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "psds-qa-e2e"
version = "0.1.0"
description = "End-to-end тестовый фреймворк для микросервисной системы видеоконсультирования"
requires-python = ">=3.10"
authors = [
  { name = "QA Team", email = "qa@example.com" },
]
dependencies = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "pytest-docker>=3.1.0",
  "pytest-xdist>=3.5.0",
  "pytest-rerunfailures>=14.0",
  "pytest-env>=1.1.0",
  "allure-pytest>=2.13.5",
  "requests>=2.31.0",
  "aiohttp>=3.9.0",
  "websockets>=12.0",
  "pydantic>=2.5.0",
  "email-validator>=2.1.0.post1",
  "faker>=22.0.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "uvloop>=0.19.0; platform_system != 'Windows'",
  "structlog>=24.0.0",
  "prometheus-client>=0.20.0",
  "mitmproxy>=11.0.0",
  "grpcio>=1.62.0",
  "grpcio-tools>=1.62.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10.0",
  "black>=24.0.0",
  "isort>=5.12.0",
  "pylint>=3.0.0",
  "coverage[toml]>=7.5.0",
  "pre-commit>=3.7.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --strict-config --alluredir=allure-results --color=yes -n auto"
testpaths = [
  "tests",
]
markers = [
  "smoke: быстрые проверки основной функциональности",
  "regression: регрессионные сценарии",
  "load: нагрузочные тесты",
  "negative: негативные сценарии",
  "websocket: тесты real-time соединений",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
show_error_codes = true
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["qa_tests"]

[tool.coverage.report]
show_missing = true
skip_covered = true

